+++ feature/community/src/main/java/com/boostcamp/dreamteam/dreamdiary/community/list/CommunityListScreen.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.dreamteam.dreamdiary.community
+package com.boostcamp.dreamteam.dreamdiary.community.list
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -12,6 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.boostcamp.dreamteam.dreamdiary.community.R
 import com.boostcamp.dreamteam.dreamdiary.community.model.DiaryUi
 import com.boostcamp.dreamteam.dreamdiary.designsystem.theme.DreamdiaryTheme
 import com.boostcamp.dreamteam.dreamdiary.ui.HomeBottomNavItem
